backend/requirements.txt
fastapi
uvicorn
requests
google-generativeai
python-dotenv
pycountry
pytest
httpx
beautifulsoup4
pandas
fpdf2
